requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
distlib==0.3.5
Django==4.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-mobile-detect==0.1
django-telegrambot==1.0.1
django-user-agents==0.4.0
filelock==3.8.0
frozenlist==1.3.1
greenlet==1.1.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
multidict==6.0.2
numpy==1.21.4
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyperclip==1.8.2
pyreadline==2.1
pyTelegramBotAPI==4.6.1
python-dateutil==1.5
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.10
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
termcolor==1.1.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
ua-parser==0.15.0
urllib3==1.26.11
user-agents==2.2.0
virtualenv==20.16.3
whitenoise==6.2.0
yarl==1.8.1
